get rid of legacy non-decoder call of item nav services
